Addresses issue raised in #trac 280 comment 8:
when you are in a filtered list, and you click on the header to add a new item, the attributes from the filter are passed onto the item create screen.

Furthermore, when you select from that screen "create another item", the attributes from the filter are propagated into subsequent created items too.