fix for #222, editing categories.
Now, the links in the table of other categories do not save any changes made to the current item, but just jump to editing the selected category.

When the update button is pressed, changes are saved, and the next category in the list is selected.  By continuing to press update on each item, you will eventually cycle through editing all of the categories in turn, and at the end return to editing the first category in the list again.

If you are creating, rather than editing, then when a newly created category is saved, the next screen allows the user to create another category.