This is the patch that differentiates the fields displayed in the item screen based on type.
Please submit comments, and requests for changes, as comments on trac #262 while that ticket is still open.  If it's been closed, open a new ticket.

If you want to have all fields available for all types (not recommended), there's a new value in config.php, forceAllFields, which you should set to TRUE.