NB this is a bleeding-edge patch.  BACKUP YOUR DATA BEFORE USING IT!
It contains a complete overhaul of the PHP for lists, because the code we had was going to be far too high a maintenance burden for 0.8 release.  Now, this new code will doubtless have bugs in it too, but it should be easier to maintain.  Specifics:
 * standardised treatments across lists and checklists
 * significant reduction in the number of lines of code handling lists
 * change in the file names of the PHP files which handle lists
 * introduction of a TYPE argument for lists (type=L) and checklists (type=C) 

The other change in this patch is to the values returned by the query function. The new values are:
 * (boolean FALSE) indicates the query failed 
 * (integer 0) query affected no rows, and returned no rows - e.g. an empty(SELECT) 
 * (integer >0) indicates the number of rows affected by an INSERT, UPDATE or DELETE 
 * (array) SELECT was successful, and has returned a number-indexed array of records, each record is an associative array of field names=>field values. 

Only that last one is unchanged: the first three are new behaviours.