Update to jQuery 1.3
This has event bubbling, so I've streamlined a lot of the event management in gtdfuncs - page startup time in listItems for long lists should be a lot quicker now.

In the calendar, parent-selector and ajax: key presses such as escape and enter should now work OK in Safari, Chrome and other Webkit browsers.

IE6 should behave a bit better when the ajax addon is enabled - layout in listItems should be ok now.